+++ lib/ex_aws/config.ex
@@ -12,72 +12,70 @@ defmodule ExAws.Config do
     :region, :security_token
   ]
 
-  def build(client, opts \\ []) do
-    config = client
-    |> ExAws.Config.get
-    |> Map.merge(Enum.into(opts, %{}))
+  def build(operation, opts \\ []) do
+    overrides = Map.new(opts)
 
-    %{client | config: config}
+    operation
+    |> ExAws.Config.get(overrides)
     |> retrieve_runtime_config
     |> parse_host_for_region
   end
 
-  def get(%{__struct__: client_module, service: service}) do
-    config_root = client_module.config_root
-    unless config_root, do: raise "A valid configuration root is required in your #{service} client"
+  @doc """
+  Builds a complete set of config for an operation.
 
-    defaults = ExAws.Config.Defaults.get
-    config   = config_root |> Keyword.get(service, [])
-    common   = defaults
-    |> Keyword.merge(config_root)
-    |> Keyword.take(@common_config)
+  1) Defaults are pulled from `ExAws.Config.Defaults`
+  2) Common values set via e.g `config :ex_aws` are merged in.
+  3) Keys set on the individual service e.g `config :ex_aws, :s3` are merged in
+  4) Finally, any configuration overrides are merged in
+  """
+  def get(op, overrides) do
+    defaults = ExAws.Config.Defaults.get(op.service)
+    common_config = Application.get_all_env(:ex_aws) |> Map.new |> Map.take(@common_config)
+    service_config = Application.get_env(:ex_aws, op.service) |> Map.new
 
     defaults
-    |> Keyword.get(service, [])
-    |> Keyword.merge(common)
-    |> Keyword.merge(config)
-    |> Enum.into(%{})
+    |> Map.merge(common_config)
+    |> Map.merge(service_config)
+    |> Map.merge(overrides)
   end
 
-  def retrieve_runtime_config(%{config: config} = client) do
-    new_config = config
-    |> Enum.reduce(%{}, fn
+  def retrieve_runtime_config(config) do
+    Enum.reduce(config, %{}, fn
       {:host, host}, config ->
-        Map.put(config, :host, retrieve_runtime_value(host, client))
+        Map.put(config, :host, retrieve_runtime_value(host, config))
       {k, v}, config ->
-        case retrieve_runtime_value(v, client) do
+        case retrieve_runtime_value(v, config) do
           %{} = result -> Map.merge(config, result)
           value -> Map.put(config, k, value)
         end
     end)
-
-    %{client | config: new_config}
   end
 
   def retrieve_runtime_value({:system, env_key}, _) do
     System.get_env(env_key)
   end
-  def retrieve_runtime_value(:instance_role, client) do
-    client
+  def retrieve_runtime_value(:instance_role, config) do
+    config
     |> ExAws.Config.AuthCache.get
     |> Map.take([:access_key_id, :secret_access_key, :security_token])
   end
-  def retrieve_runtime_value(values, client) when is_list(values) do
+  def retrieve_runtime_value(values, config) when is_list(values) do
     values
-    |> Stream.map(&retrieve_runtime_value(&1, client))
+    |> Stream.map(&retrieve_runtime_value(&1, config))
     |> Enum.find(&(&1))
   end
   def retrieve_runtime_value(value, _), do: value
 
-  def parse_host_for_region(%{config: %{host: {stub, host}, region: region} = config} = client) do
-    %{client | config: Map.put(config, :host, String.replace(host, stub, region))}
+  def parse_host_for_region(%{host: {stub, host}, region: region} = config) do
+    Map.put(config, :host, String.replace(host, stub, region))
   end
-  def parse_host_for_region(%{config: %{host: map, region: region} = config} = client) when is_map(map) do
+  def parse_host_for_region(%{host: map, region: region} = config) when is_map(map) do
     case Map.fetch(map, region) do
-      {:ok, host} -> %{client | config: Map.put(config, :host, host)}
+      {:ok, host} -> Map.put(config, :host, host)
       :error      -> "A host for region #{region} was not found in host map #{inspect(map)}"
     end
   end
-  def parse_host_for_region(client), do: client
+  def parse_host_for_region(config), do: config
 
 end
