+++ lib/ex_aws/error.ex
@@ -1,0 +1,3 @@
+defmodule ExAws.Error do
+  defexception [:message]
+end
