+++ lib/ex_aws/json/codec.ex
@@ -40,7 +40,7 @@ defmodule ExAws.JSON.Codec do
 
     def decode!(string) do
       :jsx.decode(string)
-      |> Enum.into(%{})
+      |> Map.new
     end
 
     def decode(string) do
