+++ lib/ex_aws/request.ex
@@ -12,10 +12,7 @@ defmodule ExAws.Request do
   @type error_t :: {:error, {:http_error, http_status, binary}}
   @type response_t :: success_t | error_t
 
-  def request(http_method, url, data, headers, client) do
-    config = client.config
-    service = client.service
-
+  def request(http_method, url, data, headers, config, service) do
     body = case data do
       []  -> "{}"
       d when is_binary(d) -> d
