+++ test/alternate_helper.exs
@@ -32,7 +32,7 @@ defmodule Test.JSXCodec do
 
   def decode!(string) do
     :jsx.decode(string)
-    |> Enum.into(%{})
+    |> Map.new
   end
 
   def decode(string) do
